+++ src/pyclarify/views/evaluate.py
@@ -207,6 +207,18 @@ class EvaluateParams(BaseModel):
     """
 
     items: Optional[List[ItemAggregation]] = []
+    calculations: List[Calculation]
+    data: DataQuery
+    include: List
+    format: Optional[SelectionFormat] = SelectionFormat(dataAsArray=False)
+
+
+class ExperimentalEvaluateParams(BaseModel):
+    """
+    :meta private:
+    """
+
+    items: Optional[List[ItemAggregation]] = []
     groups: Optional[List[GroupAggregation]] = []
     calculations: List[Calculation]
     data: DataQuery