+++ src/pyclarify/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Searis AS
+# Copyright 2023-2024 Searis AS
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,5 +25,5 @@ from pyclarify.views import (
 )
 import pyclarify.query
 
-__version__ = "0.6.3"
+__version__ = "0.6.4"
 __API_version__ = "1.1"